Starting refactor of strength_analysis
